+++ azext_iot/common/shared.py
@@ -214,3 +214,12 @@ class AuthenticationType(Enum):
 
     keyBased = "key"
     identityBased = "identity"
+
+
+class RenewKeyType(Enum):
+    """
+    Type of the RegenerateKey for the device.
+    """
+    Primary = "primary"
+    Secondary = "secondary"
+    Swap = "swap"